requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 评估指标
rouge-score>=0.1.2
sacrebleu>=2.3.0
jieba>=0.42.1

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 日志和监控
wandb>=0.15.0
tensorboard>=2.14.0

# Web服务 (可选)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Hugging Face Hub
huggingface-hub>=0.17.0

# 开发工具
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter支持 (可选)
jupyter>=1.0.0
ipywidgets>=8.0.0

# 其他实用工具
requests>=2.31.0
Pillow>=10.0.0 